test: update tests setup
